backend_django/requirements.txt
Django>=3.2.0
pdfminer.six>=20201018
dateparser